backend/simapp/requirements.txt
asgiref==2.3.2
async-timeout==2.0.1
attrs==18.2.0
autobahn==19.2.1
Automat==0.7.0
cachetools==3.1.0
certifi==2018.11.29
channels==2.1.7
chardet==3.0.4
constantly==15.1.0
daphne==2.2.5
defusedxml==0.5.0
Django==2.1.7
django-allauth==0.38.0
django-cors-headers==2.4.0
django-filter==2.1.0
django-js-asset==1.1.0
django-mptt==0.9.1
django-rest-auth==0.9.3
django-rest-passwordreset==0.9.7
djangorestframework==3.9.1
facebook-sdk==3.1.0
google-auth==1.6.3
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
oauthlib==3.0.0
Pillow==5.4.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
PyHamcrest==1.9.0
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
Twisted==18.9.0
txaio==18.8.1
urllib3==1.24.1
zope.interface==4.6.0
django-encrypted-filefield==0.2.1
django-encrypted-model-fields==0.5.8
celery
redis
django-celery-beat
zeep